Facilitate UI-driven analyses for non-US places (#889) (#890) (#891) (#897)
* Facilitate use of custom, non-US population URL (#890)


* Facilitate use of custom, non-US jobs url (#891)


* Facilitate use of custom, non-US OSM extracts (#899)


* fixups

* fix international jobs

+++ src/analysis/import/import_neighborhood.sh
@@ -37,6 +37,7 @@ NB_POSTGRESQL_HOST - Default: 127.0.0.1
 NB_POSTGRESQL_DB - Default: pfb
 NB_POSTGRESQL_USER - Default: gis
 NB_POSTGRESQL_PASSWORD - Default: gis
+PFB_POP_URL - Default: None
 
 "
 }
@@ -65,7 +66,9 @@ then
         usage
     else
         NB_BOUNDARY_FILE="${1}"
-        NB_STATE_FIPS="${2}"
+        NB_COUNTRY="${2}"
+        NB_STATE="${3}"
+        NB_STATE_FIPS="${4}"
 
         NB_TEMPDIR="${NB_TEMPDIR:-$(mktemp -d)}/import_neighborhood"
         mkdir -p "${NB_TEMPDIR}"
@@ -76,40 +79,49 @@ then
         update_status "IMPORTING" "Importing boundary shapefile"
         import_and_transform_shapefile "${NB_BOUNDARY_FILE}" neighborhood_boundary "${NB_INPUT_SRID}"
 
-        update_status "IMPORTING" "Downloading water blocks"
-        # Create water blocks table
-        psql -h "${NB_POSTGRESQL_HOST}" -U "${NB_POSTGRESQL_USER}" -d "${NB_POSTGRESQL_DB}" \
-            -c "
-        CREATE TABLE IF NOT EXISTS \"water_blocks\" (
-            \"STATEFP10\" integer,
-            \"COUNTYFP10\" integer,
-            \"TRACTCE10\" integer,
-            \"BLOCKCE10\" integer,
-            GEOID varchar(15),
-            \"NAME10\" char(10),
-            \"MTFCC10\" char(5),
-            \"UR10\" char(1),
-            \"UACE10\" integer,
-            \"UATYP10\" char(1),
-            \"FUNCSTAT10\" char(1),
-            \"ALAND10\" integer,
-            \"AWATER10\" bigint,
-            \"INTPTLAT10\" decimal,
-            \"INTPTLON10\" decimal
-        );"
+        if [ "${PFB_COUNTRY}" == "USA" ]; then
+            update_status "IMPORTING" "Downloading water blocks"
+            # Create water blocks table
+            psql -h "${NB_POSTGRESQL_HOST}" -U "${NB_POSTGRESQL_USER}" -d "${NB_POSTGRESQL_DB}" \
+                -c "
+            CREATE TABLE IF NOT EXISTS \"water_blocks\" (
+                \"STATEFP10\" integer,
+                \"COUNTYFP10\" integer,
+                \"TRACTCE10\" integer,
+                \"BLOCKCE10\" integer,
+                GEOID varchar(15),
+                \"NAME10\" char(10),
+                \"MTFCC10\" char(5),
+                \"UR10\" char(1),
+                \"UACE10\" integer,
+                \"UATYP10\" char(1),
+                \"FUNCSTAT10\" char(1),
+                \"ALAND10\" integer,
+                \"AWATER10\" bigint,
+                \"INTPTLAT10\" decimal,
+                \"INTPTLON10\" decimal
+            );"
 
-        # Import water file
-        WATER_FILENAME="censuswaterblocks"
-        WATER_DOWNLOAD="${NB_TEMPDIR}/${WATER_FILENAME}.zip"
-        wget -nv -O "${WATER_DOWNLOAD}" "https://s3.amazonaws.com/pfb-public-documents/censuswaterblocks.zip"
-        unzip "${WATER_DOWNLOAD}" -d "${NB_TEMPDIR}"
-        psql -h "${NB_POSTGRESQL_HOST}" -U "${NB_POSTGRESQL_USER}" -d "${NB_POSTGRESQL_DB}" \
-             -c "\copy water_blocks FROM ${NB_TEMPDIR}/${WATER_FILENAME}.csv delimiter ',' csv header"
-        echo "DONE: Importing water blocks"
+            # Import water file
+            # Only if in USA
+            WATER_FILENAME="censuswaterblocks"
+            WATER_DOWNLOAD="${NB_TEMPDIR}/${WATER_FILENAME}.zip"
+            wget -nv -O "${WATER_DOWNLOAD}" "https://s3.amazonaws.com/pfb-public-documents/censuswaterblocks.zip"
+            unzip "${WATER_DOWNLOAD}" -d "${NB_TEMPDIR}"
+            psql -h "${NB_POSTGRESQL_HOST}" -U "${NB_POSTGRESQL_USER}" -d "${NB_POSTGRESQL_DB}" \
+                -c "\copy water_blocks FROM ${NB_TEMPDIR}/${WATER_FILENAME}.csv delimiter ',' csv header"
+            echo "DONE: Importing water blocks"
+        fi
 
-        # Get blocks for the state requested
+        # Get blocks for the place requested
         update_status "IMPORTING" "Downloading census blocks"
-        NB_BLOCK_FILENAME="tabblock2010_${NB_STATE_FIPS}_pophu"
+        if [ "${PFB_COUNTRY}" == "USA" ]; then
+            NB_BLOCK_FILENAME="tabblock2010_${NB_STATE_FIPS}_pophu"
+            PFB_POP_URL="http://www2.census.gov/geo/tiger/TIGER2010BLKPOPHU/${NB_BLOCK_FILENAME}.zip"
+        else
+            NB_BLOCK_FILENAME="population"
+        fi
+
         S3_PATH="s3://${AWS_STORAGE_BUCKET_NAME}/data/${NB_BLOCK_FILENAME}.zip"
         if [ -f "/data/${NB_BLOCK_FILENAME}.zip" ]; then
             echo "Using local census blocks file"
@@ -120,13 +132,27 @@ then
             aws s3 cp "${S3_PATH}" "${BLOCK_DOWNLOAD}"
         else
             BLOCK_DOWNLOAD="${NB_TEMPDIR}/${NB_BLOCK_FILENAME}.zip"
-            wget -nv -O "${BLOCK_DOWNLOAD}" "http://www2.census.gov/geo/tiger/TIGER2010BLKPOPHU/${NB_BLOCK_FILENAME}.zip"
+            if [ "${PFB_COUNTRY}" == "USA" ]; then
+                echo "Using census blocks file from official census site"
+            else
+                echo "Using blocks file from PFB_POP_URL: ${PFB_POP_URL}"
+            fi
+            wget -nv -O "${BLOCK_DOWNLOAD}" "${PFB_POP_URL}"      
+
             if [ "${AWS_STORAGE_BUCKET_NAME}" ]; then
                 echo "Uploading census blocks file to S3 cache"
                 aws s3 cp "${BLOCK_DOWNLOAD}" "${S3_PATH}"
             fi
         fi
         unzip "${BLOCK_DOWNLOAD}" -d "${NB_TEMPDIR}"
+
+        if [ "${PFB_COUNTRY}" != "USA" ]; then
+            # Rename unzipped files if not from census so they can be found easily
+            cd $NB_TEMPDIR
+            rm "${NB_BLOCK_FILENAME}.zip"
+            for x in *; do mv "$x" "${NB_BLOCK_FILENAME}.${x##*.}"; done
+            cd - 
+        fi
 
         # Import block shapefile
         update_status "IMPORTING" "Loading census blocks"
@@ -141,13 +167,15 @@ then
                 ${NB_BOUNDARY_BUFFER});"
         echo "DONE: Finished removing blocks outside buffer"
         
-        # Discard blocks that are all water / no land area
-        update_status "IMPORTING" "Removing water blocks"
-        echo "START: Removing blocks that are 100% water from analysis"
-        psql -h "${NB_POSTGRESQL_HOST}" -U "${NB_POSTGRESQL_USER}" -d "${NB_POSTGRESQL_DB}" \
-            -c "DELETE FROM neighborhood_census_blocks AS blocks USING water_blocks \
-                AS water WHERE blocks.BLOCKID10 = water.geoid;"
-        echo "DONE: FINISHED removing blocks that are 100% water"
+        if [ "${PFB_COUNTRY}" == "USA" ]; then
+            # Discard blocks that are all water / no land area
+            update_status "IMPORTING" "Removing water blocks"
+            echo "START: Removing blocks that are 100% water from analysis"
+            psql -h "${NB_POSTGRESQL_HOST}" -U "${NB_POSTGRESQL_USER}" -d "${NB_POSTGRESQL_DB}" \
+                -c "DELETE FROM neighborhood_census_blocks AS blocks USING water_blocks \
+                    AS water WHERE blocks.BLOCKID10 = water.geoid;"
+            echo "DONE: FINISHED removing blocks that are 100% water"
+        fi
 
         BLOCK_COUNT=$(psql -h "${NB_POSTGRESQL_HOST}" -U "${NB_POSTGRESQL_USER}" -d "${NB_POSTGRESQL_DB}" \
             -t -c "SELECT count(*) as total_census_blocks FROM neighborhood_census_blocks;")
